analysis/requirements.txt
anthropic>=0.40.0
boto3>=1.34.0
Pillow>=10.0.0
